+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	layers "github.com/MGTheTrain/terraform-provider-blob-services/mgtt"
+	"github.com/MGTheTrain/terraform-provider-blob-services/mgtt"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
@@ -10,7 +10,7 @@ import (
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() *schema.Provider {
-			return layers.Provider()
+			return mgtt.Provider()
 		},
 	})
 }
